completed the `DiscordUser` struct with full spec implementation
