+++ CHANGELOG.md
@@ -11,9 +11,11 @@ New features:
 - Added a new `ActivitySpec` type which basically replaces the current architecture of a mutable `Activity` to a post-built, immutable object which can be sent around.
 - Added `Activity::empty_spec` which gives an empty but usable `ActivitySpec` (basically executes `ActivitySpec::default` internally).
 - `ActivityResponseData` now includes an `ActivityPayload` object (flattened internally) instance for you to directly see which activity details were sent.
-- New fields in `ReadyData`:
-  - `v` - RPC version (`u8`)
-  - `config` - server configuration (a new `ServerConfigurationData` object)
+- Completed the `ReadyData` struct with these changes:
+  1. New fields in `ReadyData`:
+    - `v` - RPC version (`u8`)
+    - `config` - server configuration (a new `ServerConfigurationData` object)
+  2. The inner `user` field (with type `DiscordUser`) now has the complete specification of struct fields documented in the official Discord RPC documentation.
 
 Changes:
